fix(client): prevented default max score from being less than 1 [EV-42256] (#20730)

+++ src/client/author/TestPage/components/GroupItems/PickItemCountWrapperContainer.js
@@ -45,7 +45,7 @@ export default function PickItemCountWrapperContainer({
     handleChange('deliverItemsCount', parseInt(e.target.value, 10))
   }
   const handleSectionScoreChange = (e) => {
-    handleChange('itemsDefaultMaxScore', parseInt(e.target.value, 10))
+    handleChange('itemsDefaultMaxScore', parseFloat(e.target.value, 10))
   }
   return (
     <ItemCountWrapper>
